feat: 최하단 레이아웃 1~3 단계 레이아웃에 적용 및 measurePolicy 공통화
