+++ feature-ui-create-problem/src/main/kotlin/team/duckie/app/android/feature/ui/create/problem/screen/CreateProblemScreen.kt
@@ -56,7 +56,6 @@ import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.Layout
-import androidx.compose.ui.layout.MeasurePolicy
 import androidx.compose.ui.layout.layoutId
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
@@ -73,18 +72,17 @@ import team.duckie.app.android.domain.exam.model.Answer
 import team.duckie.app.android.domain.exam.model.Question
 import team.duckie.app.android.feature.photopicker.PhotoPicker
 import team.duckie.app.android.feature.ui.create.problem.R
+import team.duckie.app.android.feature.ui.create.problem.common.CreateProblemBottomLayout
 import team.duckie.app.android.feature.ui.create.problem.common.PrevAndNextTopAppBar
+import team.duckie.app.android.feature.ui.create.problem.common.getCreateProblemMeasurePolicy
 import team.duckie.app.android.feature.ui.create.problem.viewmodel.CreateProblemViewModel
 import team.duckie.app.android.feature.ui.create.problem.viewmodel.state.CreateProblemPhotoState
 import team.duckie.app.android.feature.ui.create.problem.viewmodel.state.CreateProblemStep
 import team.duckie.app.android.util.compose.activityViewModel
-import team.duckie.app.android.util.compose.asLoose
 import team.duckie.app.android.util.compose.rememberToast
 import team.duckie.app.android.util.compose.systemBarPaddings
-import team.duckie.app.android.util.kotlin.fastFirstOrNull
 import team.duckie.app.android.util.kotlin.fastForEach
 import team.duckie.app.android.util.kotlin.fastForEachIndexed
-import team.duckie.app.android.util.kotlin.npe
 import team.duckie.quackquack.ui.border.QuackBorder
 import team.duckie.quackquack.ui.border.applyAnimatedQuackBorder
 import team.duckie.quackquack.ui.color.QuackColor
@@ -93,12 +91,10 @@ import team.duckie.quackquack.ui.component.QuackBasicTextField
 import team.duckie.quackquack.ui.component.QuackBody3
 import team.duckie.quackquack.ui.component.QuackBorderTextField
 import team.duckie.quackquack.ui.component.QuackDialog
-import team.duckie.quackquack.ui.component.QuackDivider
 import team.duckie.quackquack.ui.component.QuackDropDownCard
 import team.duckie.quackquack.ui.component.QuackImage
 import team.duckie.quackquack.ui.component.QuackRoundCheckBox
 import team.duckie.quackquack.ui.component.QuackSubtitle
-import team.duckie.quackquack.ui.component.QuackSubtitle2
 import team.duckie.quackquack.ui.icon.QuackIcon
 import team.duckie.quackquack.ui.modifier.quackClickable
 
@@ -110,52 +106,6 @@ private const val GalleryListLayoutId = "CreateProblemScreenGalleryListLayoutId"
 private const val MaximumChoice = 5
 private const val MaximumProblem = 10
 private const val TextFieldMaxLength = 20
-
-private val createProblemMeasurePolicy = MeasurePolicy { measurableItems, constraints ->
-    val looseConstraints = constraints.asLoose()
-
-    // 1. topAppBar 높이값 측정
-    val topAppBarMeasurable = measurableItems.fastFirstOrNull { measureItem ->
-        measureItem.layoutId == TopAppBarLayoutId
-    }?.measure(looseConstraints) ?: npe()
-    val topAppBarHeight = topAppBarMeasurable.height
-
-    // 2. bottomLayout, 높이값 측정
-    val bottomLayoutMeasurable = measurableItems.fastFirstOrNull { measureItem ->
-        measureItem.layoutId == BottomLayoutId
-    }?.measure(looseConstraints) ?: npe()
-    // TODO(riflockle7): 왜 이걸 더해야하는지 모르겠음.. padding 을 더하면 이렇게 됨...
-    val bottomLayoutHeight = topAppBarMeasurable.height + 72.toDp().toPx().toInt()
-
-    // 3. createProblemButton 높이값 측정
-    val contentThresholdHeight = constraints.maxHeight - topAppBarHeight - bottomLayoutHeight
-    val contentConstraints = constraints.copy(
-        minHeight = contentThresholdHeight,
-        maxHeight = contentThresholdHeight,
-    )
-    val contentMeasurable = measurableItems.fastFirstOrNull { measurable ->
-        measurable.layoutId == ContentLayoutId
-    }?.measure(contentConstraints) ?: npe()
-
-    // 3. 위에서 추출한 값들을 활용해 레이아웃 위치 처리
-    layout(
-        width = constraints.maxWidth,
-        height = constraints.maxHeight,
-    ) {
-        topAppBarMeasurable.place(
-            x = 0,
-            y = 0,
-        )
-        contentMeasurable.place(
-            x = 0,
-            y = topAppBarHeight,
-        )
-        bottomLayoutMeasurable.place(
-            x = 0,
-            y = topAppBarHeight + contentThresholdHeight,
-        )
-    }
-}
 
 /** 문제 만들기 2단계 (문제 만들기) Screen */
 @Composable
@@ -300,7 +250,11 @@ internal fun CreateProblemScreen(
             modifier = modifier
                 .fillMaxWidth()
                 .navigationBarsPadding(),
-            measurePolicy = createProblemMeasurePolicy,
+            measurePolicy = getCreateProblemMeasurePolicy(
+                TopAppBarLayoutId,
+                ContentLayoutId,
+                BottomLayoutId,
+            ),
             content = {
                 // 상단 탭바
                 PrevAndNextTopAppBar(
@@ -512,7 +466,9 @@ internal fun CreateProblemScreen(
                     modifier = Modifier
                         .fillMaxWidth()
                         .layoutId(BottomLayoutId),
-                    createButtonClick = {
+                    leftButtonLeadingIcon = QuackIcon.Plus,
+                    leftButtonText = stringResource(id = R.string.create_problem_add_problem_button),
+                    leftButtonClick = {
                         coroutineShape.launch {
                             keyboard?.hide()
                             sheetState.animateTo(ModalBottomSheetValue.Expanded)
@@ -525,7 +481,7 @@ internal fun CreateProblemScreen(
                         }
                     },
                     isMaximumProblemCount = problemCount >= MaximumProblem,
-                    isValidateCheck = vm::isValidate,
+                    isValidateCheck = vm::createProblemIsValidate,
                 )
             },
         )
@@ -1016,97 +972,5 @@ fun <T> NoLazyGridItems(
                 }
             }
         }
-    }
-}
-
-/** 문제 만들기 2단계 최하단 Layout  */
-@Composable
-fun CreateProblemBottomLayout(
-    modifier: Modifier,
-    createButtonClick: () -> Unit,
-    tempSaveButtonClick: () -> Unit,
-    nextButtonClick: () -> Unit,
-    isMaximumProblemCount: Boolean,
-    isValidateCheck: () -> Boolean,
-) {
-    val isValidate = isValidateCheck()
-    Column(modifier = modifier.background(QuackColor.White.composeColor)) {
-        QuackDivider()
-        Row(
-            modifier = Modifier.padding(horizontal = 20.dp, vertical = 12.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            Row(
-                modifier = Modifier
-                    .quackClickable {
-                        if (!isMaximumProblemCount) {
-                            createButtonClick()
-                        }
-                    }
-                    .padding(4.dp),
-            ) {
-                // TODO(riflockle7): 추후 비활성화 될 때의 resouce 이미지 필요
-                QuackImage(src = QuackIcon.Plus, size = DpSize(16.dp, 16.dp))
-                QuackSubtitle2(
-                    text = stringResource(id = R.string.create_problem_add_problem_button),
-                    color = if (isMaximumProblemCount) QuackColor.Gray2 else QuackColor.Black,
-                )
-            }
-
-            Spacer(modifier = Modifier.weight(1f))
-
-            // 임시저장 버튼
-            QuackSubtitle(
-                modifier = Modifier
-                    .clip(RoundedCornerShape(size = 8.dp))
-                    .background(QuackColor.White.composeColor)
-                    .quackClickable { tempSaveButtonClick() }
-                    .applyAnimatedQuackBorder(
-                        QuackBorder(1.dp, QuackColor.Gray3),
-                        shape = RoundedCornerShape(size = 8.dp),
-                    )
-                    .padding(vertical = 12.dp, horizontal = 19.dp),
-                color = QuackColor.Black,
-                text = stringResource(id = R.string.create_problem_temp_save_button),
-            )
-
-            // 다음 버튼
-            QuackSubtitle(
-                modifier = Modifier
-                    .clip(RoundedCornerShape(size = 8.dp))
-                    .background(
-                        if (isValidate) {
-                            QuackColor.DuckieOrange.composeColor
-                        } else {
-                            QuackColor.Gray2.composeColor
-                        }
-                    )
-                    .quackClickable {
-                        if (isValidate) {
-                            nextButtonClick()
-                        }
-                    }
-                    .applyAnimatedQuackBorder(
-                        QuackBorder(
-                            1.dp,
-                            if (isValidate) {
-                                QuackColor.DuckieOrange
-                            } else {
-                                QuackColor.Gray2
-                            }
-                        ),
-                        shape = RoundedCornerShape(size = 8.dp),
-                    )
-                    .padding(
-                        vertical = 12.dp,
-                        horizontal = 19.dp,
-                    ),
-                color = QuackColor.White,
-                text = stringResource(id = R.string.next),
-            )
-        }
-
-        QuackDivider()
     }
 }
